% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NtM.R
\name{scrape_cleaner}
\alias{scrape_cleaner}
\title{Area matcher from scrapes}
\usage{
scrape_cleaner(.scrape_df, AOR_search, ...)
}
\description{
Area matcher from scrapes
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
