% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NtM.R
\name{scrape_NtM}
\alias{scrape_NtM}
\title{Yearly Note To Mariners Scrapes}
\usage{
scrape_NtM(year, format)
}
\description{
Yearly Note To Mariners Scrapes
}
\author{
Brendan Knapp, \email{brendan.knapp@nps.edu}

Christopher Callaghan, \email{cjcallag@nps.edu}
}
